updated regarding to comments
