+++ classes/app/src/main/res/layout/activity_login.xml
@@ -54,7 +54,7 @@
                     app:layout_constraintStart_toStartOf="parent"/>
 
             <Button
-                    android:id="@+id/signIn"
+                    android:id="@+id/signInButton"
                     android:text="@string/sign_in"
                     style="@style/LoginScreenButton"
                     app:layout_constraintTop_toBottomOf="@id/password"
@@ -65,13 +65,13 @@
                     android:id="@+id/forgotPassword"
                     style="@style/LoginScreenForgotPassword"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toBottomOf="@id/signIn"
+                    app:layout_constraintTop_toBottomOf="@id/signInButton"
                     app:layout_constraintEnd_toEndOf="parent"
                     android:text="@string/forgot_password"
                     android:layout_marginBottom="@dimen/dimen_large"/>
 
             <Button
-                    android:id="@+id/signUp"
+                    android:id="@+id/signUpButton"
                     android:backgroundTint="@color/signUpButton"
                     android:text="@string/sign_up"
                     style="@style/LoginScreenButton"
